stream.hls: move HLSSegment to dedicated module
