+++ src/streamlink/stream/hls/__init__.py
@@ -1,6 +1,11 @@
 from streamlink.stream.hls.hls import HLSStream, HLSStreamReader, HLSStreamWorker, HLSStreamWriter, MuxedHLSStream
 from streamlink.stream.hls.m3u8 import (
     M3U8,
+    M3U8Parser,
+    parse_m3u8,
+    parse_tag,
+)
+from streamlink.stream.hls.segment import (
     ByteRange,
     DateRange,
     ExtInf,
@@ -8,12 +13,9 @@ from streamlink.stream.hls.m3u8 import (
     HLSSegment,
     IFrameStreamInfo,
     Key,
-    M3U8Parser,
     Map,
     Media,
     Resolution,
     Start,
     StreamInfo,
-    parse_m3u8,
-    parse_tag,
 )
